+++ Sources/RichTextKit/Component/RichTextViewComponent+Indent.swift
@@ -10,9 +10,7 @@ import Foundation
 
 #if canImport(UIKit)
 import UIKit
-#endif
-
-#if canImport(AppKit) && !targetEnvironment(macCatalyst)
+#elseif canImport(AppKit) && !targetEnvironment(macCatalyst)
 import AppKit
 #endif
 
@@ -20,22 +18,16 @@ public extension RichTextViewComponent {
 
     /// Get the rich text indent.
     var richTextIndent: CGFloat? {
-        richTextParagraphStyle?.headIndent
+        richTextParagraphStyleValue(\.headIndent)
     }
 
     /// Set the rich text indent.
     func setRichTextIndent(to val: CGFloat) {
-        let style = NSMutableParagraphStyle(
-            from: richTextParagraphStyle,
-            indent: val
-        )
-        setRichTextParagraphStyle(style)
+        setRichTextParagraphStyleValue(\.headIndent, val)
     }
 
     /// Step the rich text indent.
     func stepRichTextIndent(points: CGFloat) {
-        let old = richTextParagraphStyle?.headIndent ?? 0
-        let new = max(0, old + points)
-        setRichTextIndent(to: new)
+        stepRichTextParagraphStyleValue(\.headIndent, points)
     }
 }